+++ client/src/main/scala/org/http4s/client/JavaNetClientBuilder.scala
@@ -1,14 +1,29 @@
 /*
- * Copyright 2013-2020 http4s.org
+ * Copyright 2014 http4s.org
  *
- * SPDX-License-Identifier: Apache-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.http4s
 package client
 
+<<<<<<< HEAD
 import cats.effect.{Async, Resource, Sync}
 import cats.implicits._
+=======
+import cats.effect.{Async, Blocker, ContextShift, Resource, Sync}
+import cats.syntax.all._
+>>>>>>> cats-effect-3
 import fs2.Stream
 import fs2.io.{readInputStream, writeOutputStream}
 import java.io.IOException
@@ -96,8 +111,15 @@ sealed abstract class JavaNetClientBuilder[F[_]] private (
           _ <- F.delay(conn.setConnectTimeout(timeoutMillis(connectTimeout)))
           _ <- F.delay(conn.setReadTimeout(timeoutMillis(readTimeout)))
           _ <- F.delay(conn.setRequestMethod(req.method.renderString))
-          _ <- F.delay(req.headers.foreach { case Header(name, value) =>
-            conn.setRequestProperty(name.toString, value)
+          _ <- F.delay(req.headers.foreach {
+            case Header(name, value) =>
+              conn.setRequestProperty(name.toString, value)
+            case h: Header.Parsed =>
+              // Appease 2.13.4's exhaustiveness checker
+              conn.setRequestProperty(h.name.toString, h.value)
+            case Header.Raw(name, value) =>
+              // Appease 2.13.4's exhaustiveness checker
+              conn.setRequestProperty(name.toString, value)
           })
           _ <- F.delay(conn.setInstanceFollowRedirects(false))
           _ <- F.delay(conn.setDoInput(true))