+++ core/src/main/scala/org/http4s/parser/LinkHeader.scala
@@ -1,7 +1,17 @@
 /*
- * Copyright 2013-2020 http4s.org
+ * Copyright 2013 http4s.org
  *
- * SPDX-License-Identifier: Apache-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.http4s
@@ -46,7 +56,11 @@ private[parser] trait LinkHeader {
     def LinkValueAttr: Rule[LinkValue :: HNil, LinkValue :: HNil] =
       rule {
         "rel=" ~ (Token | QuotedString) ~> { (link: LinkValue, rel: String) =>
-          link.copy(rel = Some(rel))
+          // https://tools.ietf.org/html/rfc8288#section-3.3
+          if (link.rel.isDefined)
+            link
+          else
+            link.copy(rel = Some(rel))
         } |
           "rev=" ~ (Token | QuotedString) ~> { (link: LinkValue, rev: String) =>
             link.copy(rev = Some(rev))