+++ dropwizard-metrics/src/test/scala/org/http4s/metrics/dropwizard/DropwizardServerMetricsSpec.scala
@@ -1,7 +1,17 @@
 /*
- * Copyright 2013-2020 http4s.org
+ * Copyright 2018 http4s.org
  *
- * SPDX-License-Identifier: Apache-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.http4s.metrics.dropwizard
@@ -157,7 +167,7 @@ class DropwizardServerMetricsSpec extends Http4sSpec with Http4sLegacyMatchersIO
       count(registry, Counter("server.default.active-requests")) must beEqualTo(0L)
       count(registry, Timer("server.default.requests.total")) must beEqualTo(1L)
       valuesOf(registry, Timer("server.default.requests.headers")) must beSome(Array(50000000L))
-      valuesOf(registry, Timer("server.default.get-requests")) must beSome(Array(100000000L))
+      valuesOf(registry, Timer("server.default.get-requests")) must beSome(Array(50000000L))
     }
 
     "register an abnormal termination" in {