+++ server/src/main/scala/org/http4s/server/middleware/RequestId.scala
@@ -1,7 +1,17 @@
 /*
- * Copyright 2013-2020 http4s.org
+ * Copyright 2014 http4s.org
  *
- * SPDX-License-Identifier: Apache-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.http4s
@@ -13,8 +23,13 @@ import org.http4s.{Header, Http, Request, Response}
 import cats.{FlatMap, ~>}
 import cats.arrow.FunctionK
 import cats.data.{Kleisli, OptionT}
+<<<<<<< HEAD
 import cats.effect.{Sync, SyncIO}
 import cats.implicits._
+=======
+import cats.effect.{IO, Sync}
+import cats.syntax.all._
+>>>>>>> cats-effect-3
 import org.typelevel.ci.CIString
 import io.chrisdavenport.vault.Key
 import java.util.UUID