+++ server/src/main/scala/org/http4s/server/middleware/authentication/DigestAuth.scala
@@ -1,7 +1,17 @@
 /*
- * Copyright 2013-2020 http4s.org
+ * Copyright 2014 http4s.org
  *
- * SPDX-License-Identifier: Apache-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.http4s
@@ -12,7 +22,7 @@ package authentication
 import cats.Applicative
 import cats.data.{Kleisli, NonEmptyList}
 import cats.effect.Sync
-import cats.implicits._
+import cats.syntax.all._
 import java.math.BigInteger
 import java.security.SecureRandom
 import java.util.Date
@@ -85,7 +95,7 @@ object DigestAuth {
     req.headers.get(Authorization) match {
       case Some(Authorization(Credentials.AuthParams(AuthScheme.Digest, params))) =>
         checkAuthParams(realm, store, nonceKeeper, req, params)
-      case Some(Authorization(_)) =>
+      case Some(_) =>
         F.pure(NoCredentials)
       case None =>
         F.pure(NoAuthorizationHeader)