+++ testing/src/test/scala/org/http4s/Http4sSpec.scala
@@ -11,7 +11,7 @@
 package org.http4s
 
 import cats.effect.{IO, Resource}
-import cats.implicits._
+import cats.syntax.all._
 import fs2._
 import fs2.text._
 import java.util.concurrent.{ScheduledExecutorService, ScheduledThreadPoolExecutor, TimeUnit}